+++ HISTORY.rst
@@ -7,6 +7,9 @@ History
 * Drop official support for Python 3.4 (removing from test matrix, code may still work).
 * Add support for Python 3.7 and 3.8 (adding to test matrix).
 * Update development requirements.
+* State machine names should now be fully qualified for mixins, simple names are deprecated and
+  will no longer be supported on a future version.
+* Development: Adding mypy linter.
 
 
 0.7.1 (2019-01-18)